% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6BinSeg.R
\name{Cost_L2}
\alias{Cost_L2}
\title{Cost_L2 cost module}
\description{
Computes the L2 cost (squared error) for a segment of a matrix.
}
\details{
The L2 cost is the sum of squared deviations from the mean.
}
