% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6BinSeg.R
\name{RCostClass}
\alias{RCostClass}
\title{R-based custom cost module}
\description{
Provides a bridge for user-supplied cost functions implemented in R.
}
\details{
Allows flexible experimentation with any R-based cost function.
}
